feat(consensus): Change meta.score and meta.accumulated_weight to int type

+++ hathor/event/model/event_data.py
@@ -45,8 +45,8 @@ class TxMetadata(BaseModel, extra=Extra.ignore):
     received_by: List[int]
     children: List[str]
     twins: List[str]
-    accumulated_weight: float
-    score: float
+    accumulated_weight: int
+    score: int
     first_block: Optional[str]
     height: int
     validation: str
